refactor(trusty-search): split mcp/tools.rs into per-tool-group submodules (#607 PR-B)
Split the 2,187-SLOC monolith crates/trusty-search/src/mcp/tools.rs into a
tools/ module directory — each file is ≤ 500 lines:

- mod.rs (279)       — thin re-export facade + McpServer::dispatch + call_tool
- descriptors.rs (348) — static tool_descriptors() for tools/list
- http.rs (153)      — shared HTTP transport helpers (get/post/delete)
- index.rs (146)     — dispatch_index_tool: index_file, list_indexes, …
- misc.rs (166)      — dispatch_misc_tool: grep, chat, search_health, upgrade
- search.rs (455)    — dispatch_search_tool + per-lane routing + SearchLane
- types.rs (132)     — DispatchError, require_str, response-wrapping helpers
- tests.rs (407)     — core protocol + spawn_mock_daemon shared helper
- tests_lane.rs (328) — per-lane tool tests (issue #138)
- tests_tools_list.rs (169) — tools/list completeness + summarise_stages tests
- tests_empty_query.rs (337) — empty-query rejection tests (issue #882)

Public API is unchanged; all call sites continue to use
`crate::mcp::tools::{McpServer, tool_descriptors, error_codes, …}`.

Removes the 2,187-line allowlist entry; all new files are under the 500-line cap.
PR-B of the 7-PR #607 line-cap split program.
